+++ packages/triple-document/src/index.tsx
@@ -1,5 +1,6 @@
 import { TripleDocument } from './triple-document'
 export { ELEMENTS } from './triple-document'
+export { TripleElementData as TripleDocumentElementData } from './types'
 export * from './text'
 export * from './tna'
 export * from './region'